refactor(TableReel): useReelCellsをTableReelに統合
