+++ packages/smarthr-ui/src/components/Table/TableReel.tsx
@@ -1,10 +1,22 @@
 'use client'
 
-import { type ComponentPropsWithRef, type FC, type PropsWithChildren, useMemo } from 'react'
+import {
+  type ComponentPropsWithRef,
+  type FC,
+  type PropsWithChildren,
+  useEffect,
+  useMemo,
+  useState,
+} from 'react'
 import { tv } from 'tailwind-variants'
 
 import { reelShadowClassNameGenerator } from './reelShadowStyle'
-import { useReelCells } from './useReelCells'
+
+const TR_SELECTOR = 'table tr'
+const FIXED_LEFT_SELECTOR = '[data-fixed="left"]'
+const FIXED_RIGHT_SELECTOR = '[data-fixed="right"]'
+
+const HAS_FIXED_SELECTOR = `${TR_SELECTOR} ${FIXED_LEFT_SELECTOR},${TR_SELECTOR} ${FIXED_RIGHT_SELECTOR}`
 
 type Props = PropsWithChildren &
   Omit<ComponentPropsWithRef<'div'>, keyof PropsWithChildren> & {
@@ -19,7 +31,104 @@ const classNameGenerator = tv({
 })
 
 export const TableReel: FC<Props> = ({ className, children, tableWrapperRef, ...rest }) => {
-  const { showShadow } = useReelCells(tableWrapperRef)
+  const [showShadow, setShowShadow] = useState(false)
+
+  useEffect(() => {
+    const wrapper = tableWrapperRef.current
+
+    if (!wrapper) {
+      return
+    }
+
+    let cellCleans: Array<() => void> = []
+    const allClean = () => {
+      cellCleans.forEach((clean) => clean())
+      cellCleans = []
+    }
+
+    const handleScroll = () => {
+      if (!wrapper.querySelector(HAS_FIXED_SELECTOR)) {
+        return
+      }
+
+      let isVisible = false
+      const commonAction = (
+        cells: HTMLElement[] | NodeListOf<HTMLElement>,
+        direction: 'left' | 'right',
+        visible: boolean,
+      ) => {
+        allClean()
+
+        const action = () => {
+          let position = 0
+
+          cells.forEach((cell, index) => {
+            if (cell.classList.toggle('fixed', visible)) {
+              isVisible = true
+              cell.style[direction] = `${position}px`
+              cell.style.zIndex = (index + 1).toString()
+
+              position += cell.offsetWidth
+            }
+          })
+        }
+
+        action()
+        const observer = new ResizeObserver(action)
+
+        cells.forEach((cell) => {
+          observer.observe(cell)
+        })
+
+        cellCleans.push(() => {
+          cells.forEach((cell) => {
+            observer.unobserve(cell)
+          })
+        })
+      }
+
+      wrapper.querySelectorAll<HTMLElement>(TR_SELECTOR).forEach((tr) => {
+        const leftCells = tr.querySelectorAll<HTMLElement>(FIXED_LEFT_SELECTOR)
+        const rightCells = Array.from(
+          tr.querySelectorAll<HTMLElement>(FIXED_RIGHT_SELECTOR),
+        ).reverse()
+
+        if (leftCells.length > 0) {
+          commonAction(leftCells, 'left' as const, wrapper.scrollLeft > 0)
+        }
+
+        if (rightCells.length > 0) {
+          commonAction(
+            rightCells,
+            'right' as const,
+            wrapper.scrollLeft < wrapper.scrollWidth - wrapper.clientWidth - 1,
+          )
+        }
+      })
+
+      setShowShadow(isVisible)
+    }
+
+    handleScroll()
+    wrapper.addEventListener('scroll', handleScroll)
+
+    const resizeObserver = new ResizeObserver(handleScroll)
+    resizeObserver.observe(wrapper)
+
+    // HINT: Paginationと組み合わせた際などにテーブル構造の変更を検知して再生成
+    const mutationObserver = new MutationObserver(handleScroll)
+    mutationObserver.observe(wrapper, {
+      childList: true,
+      subtree: true,
+    })
+
+    return () => {
+      wrapper.removeEventListener('scroll', handleScroll)
+      resizeObserver.unobserve(wrapper)
+      mutationObserver.disconnect()
+      allClean()
+    }
+  }, [tableWrapperRef])
 
   const classNames = useMemo(() => {
     const { wrapper, inner } = classNameGenerator()
